refactor(api): move new certification version route and his controller to certification-version-route.js
