+++ api/src/certification/configuration/application/certification-framework-controller.js
@@ -17,22 +17,6 @@ const getFrameworkHistory = async function (request) {
   return frameworkHistorySerializer.serialize({ scope, frameworkHistory });
 };
 
-const createCertificationVersion = async function (request, h) {
-  const { scope } = request.params;
-  const { tubeIds } = request.payload.data.attributes;
-
-  await usecases.createCertificationVersion({ scope, tubeIds });
-
-  return h
-    .response({
-      data: {
-        id: scope,
-        type: 'certification-consolidated-framework',
-      },
-    })
-    .code(201);
-};
-
 const getTargetProfileHistory = async function (request) {
   const scope = request.params.scope;
   const certificationFramework = await usecases.getComplementaryCertificationTargetProfileHistory({
@@ -42,7 +26,6 @@ const getTargetProfileHistory = async function (request) {
 };
 
 const certificationFrameworkController = {
-  createCertificationVersion,
   findCertificationFrameworks,
   getFrameworkHistory,
   getTargetProfileHistory,
