+++ api/src/certification/configuration/application/certification-framework-route.js
@@ -1,7 +1,6 @@
 import Joi from 'joi';
 
 import { securityPreHandlers } from '../../../shared/application/security-pre-handlers.js';
-import { identifiersType } from '../../../shared/domain/types/identifiers-type.js';
 import { SCOPES } from '../../shared/domain/models/Scopes.js';
 import { certificationFrameworkController } from './certification-framework-controller.js';
 
@@ -90,42 +89,6 @@ const register = async function (server) {
         notes: [
           'Cette route est restreinte aux utilisateurs authentifiés avec le rôle Super Admin, Support, Certif et Métier',
           "Elle renvoie l'historique des profils cibles rattachés à un référentiel de certification.",
-        ],
-      },
-    },
-    {
-      method: 'POST',
-      path: '/api/admin/frameworks/{scope}/new-version',
-      config: {
-        pre: [
-          {
-            method: (request, h) =>
-              securityPreHandlers.hasAtLeastOneAccessOf([securityPreHandlers.checkAdminMemberHasRoleSuperAdmin])(
-                request,
-                h,
-              ),
-            assign: 'hasRoleSuperAdmin',
-          },
-        ],
-        validate: {
-          params: Joi.object({
-            scope: Joi.string()
-              .required()
-              .valid(...Object.values(SCOPES)),
-          }),
-          payload: Joi.object({
-            data: {
-              attributes: {
-                tubeIds: Joi.array().items(identifiersType.tubeId).min(1).unique().required(),
-              },
-            },
-          }),
-        },
-        handler: certificationFrameworkController.createCertificationVersion,
-        tags: ['api', 'admin'],
-        notes: [
-          'Cette route est restreinte aux utilisateurs authentifiés avec le rôle Super Admin',
-          'Elle permet de créer une nouvelle version de référentiel de certification',
         ],
       },
     },
