+++ api/tests/certification/configuration/unit/application/certification-framework-route_test.js
@@ -154,46 +154,4 @@ describe('Unit | Certification | Configuration | Application | Router | certific
       });
     });
   });
-
-  describe('POST /api/admin/frameworks/{scope}/new-version', function () {
-    describe('when the user authenticated has no role', function () {
-      it('should return 403 HTTP status code', async function () {
-        // given
-        sinon
-          .stub(securityPreHandlers, 'hasAtLeastOneAccessOf')
-          .returns((request, h) => h.response().code(403).takeover());
-        sinon.stub(certificationFrameworkController, 'createCertificationVersion').returns('ok');
-        const httpTestServer = new HttpTestServer();
-        await httpTestServer.register(moduleUnderTest);
-
-        // when
-        const response = await httpTestServer.request('POST', `/api/admin/frameworks/${SCOPES.CORE}/new-version`, {
-          data: { attributes: { tubeIds: ['tubeId'] } },
-        });
-
-        // then
-        expect(response.statusCode).to.equal(403);
-        sinon.assert.notCalled(certificationFrameworkController.createCertificationVersion);
-      });
-    });
-
-    describe('when the scope is invalid', function () {
-      it('should return 400 HTTP status code', async function () {
-        // given
-        sinon.stub(securityPreHandlers, 'hasAtLeastOneAccessOf').callsFake(() => () => true);
-        sinon.stub(certificationFrameworkController, 'createCertificationVersion').returns('ok');
-        const httpTestServer = new HttpTestServer();
-        await httpTestServer.register(moduleUnderTest);
-
-        // when
-        const response = await httpTestServer.request('POST', '/api/admin/frameworks/INVALID_SCOPE/new-version', {
-          data: { attributes: { tubeIds: ['tubeId'] } },
-        });
-
-        // then
-        expect(response.statusCode).to.equal(400);
-        sinon.assert.notCalled(certificationFrameworkController.createCertificationVersion);
-      });
-    });
-  });
 });
