+++ api/tests/certification/configuration/unit/application/certification-version-route_test.js
@@ -2,6 +2,7 @@ import sinon from 'sinon';
 
 import { certificationVersionController } from '../../../../../src/certification/configuration/application/certification-version-controller.js';
 import * as moduleUnderTest from '../../../../../src/certification/configuration/application/certification-version-route.js';
+import { SCOPES } from '../../../../../src/certification/shared/domain/models/Scopes.js';
 import { securityPreHandlers } from '../../../../../src/shared/application/security-pre-handlers.js';
 import { expect } from '../../../../test-helper.js';
 import { HttpTestServer } from '../../../../tooling/server/http-test-server.js';
@@ -192,6 +193,48 @@ describe('Unit | Certification | Configuration | Application | Router | certific
       // then
       expect(response.statusCode).to.equal(400);
       sinon.assert.notCalled(certificationVersionController.deleteCertificationVersion);
+    });
+  });
+
+  describe('POST /api/admin/frameworks/{scope}/new-version', function () {
+    describe('when the user authenticated has no role', function () {
+      it('should return 403 HTTP status code', async function () {
+        // given
+        sinon
+          .stub(securityPreHandlers, 'hasAtLeastOneAccessOf')
+          .returns((request, h) => h.response().code(403).takeover());
+        sinon.stub(certificationVersionController, 'createCertificationVersion').returns('ok');
+        const httpTestServer = new HttpTestServer();
+        await httpTestServer.register(moduleUnderTest);
+
+        // when
+        const response = await httpTestServer.request('POST', `/api/admin/frameworks/${SCOPES.CORE}/new-version`, {
+          data: { attributes: { tubeIds: ['tubeId'] } },
+        });
+
+        // then
+        expect(response.statusCode).to.equal(403);
+        sinon.assert.notCalled(certificationVersionController.createCertificationVersion);
+      });
+    });
+
+    describe('when the scope is invalid', function () {
+      it('should return 400 HTTP status code', async function () {
+        // given
+        sinon.stub(securityPreHandlers, 'hasAtLeastOneAccessOf').callsFake(() => () => true);
+        sinon.stub(certificationVersionController, 'createCertificationVersion').returns('ok');
+        const httpTestServer = new HttpTestServer();
+        await httpTestServer.register(moduleUnderTest);
+
+        // when
+        const response = await httpTestServer.request('POST', '/api/admin/frameworks/INVALID_SCOPE/new-version', {
+          data: { attributes: { tubeIds: ['tubeId'] } },
+        });
+
+        // then
+        expect(response.statusCode).to.equal(400);
+        sinon.assert.notCalled(certificationVersionController.createCertificationVersion);
+      });
     });
   });
 });
